requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
djangorestframework>=3.15
django-filter>=24.3
djangorestframework-simplejwt>=5.3
